requirements.txt
gradio
llama-cpp-python==0.2.90
huggingface-hub
requests
